+++ src/api/utils/mod.rs
@@ -3,3 +3,4 @@ mod validation_error_to_http;
 
 pub use json_responce::json_responce;
 pub use validation_error_to_http::validation_error_to_http;
+pub use validation_error_to_http::ValidationError;
